v3.1.2

+++ mod/output/report/general.py
@@ -1,5 +1,6 @@
 # -*- coding:utf-8 -*-
 
+import os
 from mod.tools.check import Check
 from mod.tools.debug import Debug
 from mod.tools.io_tools import write_to_html, delete_directory
@@ -67,4 +68,10 @@ def archive_to_report(Queue_Output, ruleldict, input_argv, unarchive_path):
     # 将结果写入到 html 文件中
     write_to_html(finish_data, input_argv)
     # 清除临时目录
-    delete_directory(unarchive_path)
+    temp_path = os.path.join(os.path.abspath(os.path.join(os.path.realpath(__file__), '..\..\..\..')), Check.get_temp_path())
+    # 代表分析的是压缩包，需要清空临时目录
+    if temp_path == unarchive_path:
+        delete_directory(unarchive_path)
+    # 代表分析的是单独的文件，不需要执行此步骤
+    else:
+        pass
